test(ci): split managed image retention coverage
